ContriBlock-Major-Project-2025/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
web3==6.20.1
py-solc-x==2.0.3
requests==2.32.3
alembic==1.13.2
bcrypt==4.2.0

